# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TEQUALY Field Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /obras:
    get:
      operationId: listObras
      tags: [obras]
      summary: List all obras
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of obras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obra"
    post:
      operationId: createObra
      tags: [obras]
      summary: Create a new obra
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObraInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"
  /obras/{id}:
    get:
      operationId: getObra
      tags: [obras]
      summary: Get obra by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Obra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"
        "404":
          description: Not found
    patch:
      operationId: updateObra
      tags: [obras]
      summary: Update an obra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObraUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obra"
    delete:
      operationId: deleteObra
      tags: [obras]
      summary: Delete an obra
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /obras/{id}/dashboard:
    get:
      operationId: getObraDashboard
      tags: [obras]
      summary: Per-obra dashboard summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObraDashboard"

  # ─── Contratos ───────────────────────────────────────────
...
components:
  schemas:
...
    ActivityItem:
      type: object
      required: [id, tipo, descricao, entidade, entidade_id, created_at]
      properties:
        id:
          type: integer
        tipo:
          type: string
        descricao:
          type: string
        entidade:
          type: string
        entidade_id:
          type: integer
        usuario:
          type: ["string", "null"]
        created_at:
          type: string
    Obra:
      type: object
      required: [id, nome, codigo, status, created_at]
      properties:
        id:
          type: integer
        nome:
          type: string
        codigo:
          type: string
        descricao:
          type: ["string", "null"]
        status:
          type: string
        data_inicio:
          type: ["string", "null"]
        data_fim_prevista:
          type: ["string", "null"]
        data_fim_real:
          type: ["string", "null"]
        localizacao:
          type: ["string", "null"]
        gestor_responsavel:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    ObraInput:
      type: object
      required: [nome, codigo, status]
      properties:
        nome:
          type: string
          minLength: 1
        codigo:
          type: string
          minLength: 1
        descricao:
          type: string
        status:
          type: string
        data_inicio:
          type: string
        data_fim_prevista:
          type: string
        localizacao:
          type: string
        gestor_responsavel:
          type: string
    ObraUpdate:
      type: object
      properties:
        nome:
          type: string
        codigo:
          type: string
        descricao:
          type: string
        status:
          type: string
        data_inicio:
          type: string
        data_fim_prevista:
          type: string
        data_fim_real:
          type: string
        localizacao:
          type: string
        gestor_responsavel:
          type: string
    ObraDashboard:
      type: object
      required: [obra, totalVisitas, rncsAbertos, compromissosVencidos, totalAtivos, totalFrentes, totalContratos]
      properties:
        obra:
          $ref: "#/components/schemas/Obra"
        totalVisitas:
          type: integer
        rncsAbertos:
          type: integer
        compromissosVencidos:
          type: integer
        totalAtivos:
          type: integer
        totalFrentes:
          type: integer
        totalContratos:
          type: integer
        kpis:
          type: array
          items:
            $ref: "#/components/schemas/Kpi"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
...
    Kpi:
      type: object
      required: [id, obra_id, nome, valor, unidade, periodo, created_at]
      properties:
        id:
          type: integer
        obra_id:
          type: integer
        nome:
          type: string
        valor:
          type: number
        unidade:
          type: string
        periodo:
          type: string
        meta:
          type: ["number", "null"]
        created_at:
          type: string
...